+++ Mailer/migrations/20210528085728_add_unique_index_to_mail_type_code_column.php
@@ -1,0 +1,14 @@
+<?php
+declare(strict_types=1);
+
+use Phinx\Migration\AbstractMigration;
+
+final class AddUniqueIndexToMailTypeCodeColumn extends AbstractMigration
+{
+    public function change(): void
+    {
+        $this->table('mail_types')
+            ->addIndex('code', ['unique' => true])
+            ->update();
+    }
+}